+++ test/store.test.ts
@@ -204,6 +204,50 @@ describe('connect-mssql-v2', () => {
         return done();
       });
     });
+
+    describe('sessionError Listener', () => {
+      test('Should emit sessionError (errorHandler called directly)', (done) => {
+        const store = new MSSQLStore(sqlConfig, {
+          table: 'Sessions',
+        });
+
+        store.on('sessionError', (error: any, method: string) => {
+          expect(method).toEqual('test');
+          expect(error).toBeInstanceOf(Error);
+
+          done();
+        });
+
+        const errorHandler = store.errorHandler(
+          'test',
+          new Error('Test errorHandler'),
+          () => true,
+        );
+
+        expect(errorHandler).toBeTruthy();
+      });
+
+      test('Should emit sessionError (errorHandler called directly, NO CB)', (done) => {
+        const store = new MSSQLStore(sqlConfig, {
+          table: 'Sessions',
+        });
+
+        store.on('sessionError', (error: any, method: string) => {
+          expect(method).toEqual('test');
+          expect(error).toBeInstanceOf(Error);
+
+          done();
+        });
+
+        const errorHandler = store.errorHandler(
+          'test',
+          new Error('Test errorHandler'),
+        );
+
+        expect(errorHandler).toBeFalsy();
+      });
+    });
+
     test('Should emit an error listener', (done) => {
       const localConfig = {
         user: process.env.SQLUSER,